pyproject.toml
[tool.poetry]
name = "pypeprompts"
version = "0.0.2"
description = "A simple tracker for OpenAI API calls for Pype App"
authors = ["dhruv <dhruv@pypeai.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pypeprompts"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = "^2.25.1"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
responses = "^0.25.3"
pytest-asyncio = "^0.23.8"
aiohttp = "^3.10.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
openai = "^1.40.1"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]